==> 01_pre_processing.ipynb <==
import pandas as pd

from src.cloud import download
from src.crypto import decrypt_files
from src.file_handling import process_folder
from src.calibration_file_processing import process_calibration_file
from src.spiral_file_processing import process_spiral_accelerometer_file, process_spiral_drawing_file
from src.stroop_file_processing import process_stroop_file
from src.drift_file_processing import process_drift_file
from src.slow_move_file_processing import process_slow_move_accelerations_file, process_slow_move_touches_file

local_folder = './cloud'
local_encrypted_folder = local_folder+'/encrypted'
local_decrypted_folder = local_folder+'/decrypted'

download(local_encrypted_folder)

# Runs for ~ 217 minutes
decrypt_files(local_encrypted_folder, local_decrypted_folder)

def participants_with_less_sessions_than_expected(df, expected=10):
    session_count_per_participant = df.groupby(['subject','uuid'])[['uuid']].agg('first').groupby(['subject']).agg('count')
    return session_count_per_participant[session_count_per_participant['uuid']< expected]

# Calibration

calibration_folder = local_decrypted_folder+'/calibration'
calibrations = process_folder(calibration_folder, process_calibration_file)
calibrations_df = pd.DataFrame(calibrations)
calibrations_df.to_csv('calibrations.csv', index=False)

participants_with_less_sessions_than_expected(calibrations_df, expected=5)

# Spiral Drawing

spiral_folder = local_decrypted_folder+'/spiral'
spiral_accelerations = process_folder(spiral_folder, process_spiral_accelerometer_file)
spiral_accelerations_df = pd.DataFrame(spiral_accelerations)
spiral_accelerations_df.to_csv('spiral_accelerations.csv', index=False)

spiral_drawings = process_folder(spiral_folder, process_spiral_drawing_file)
spiral_drawings_df = pd.DataFrame(spiral_drawings)
spiral_drawings_df.to_csv('spiral_drawings.csv', index=False)

participants_with_less_sessions_than_expected(spiral_accelerations_df)

# Stroop Game

stroop_folder = local_decrypted_folder+'/static'
stroop_sessions = process_folder(stroop_folder, process_stroop_file)
stroop_sessions_df = pd.DataFrame(stroop_sessions)
stroop_sessions_df.to_csv('stroop_accelerations.csv', index=False)

participants_with_less_sessions_than_expected(stroop_sessions_df)

# Drift Test

drift_folder = local_decrypted_folder+'/drift'
drift_tests = process_folder(drift_folder, process_drift_file)
drift_tests_df = pd.DataFrame(drift_tests)
drift_tests_df.to_csv('drift_accelerations.csv', index=False)

participants_with_less_sessions_than_expected(drift_tests_df, expected=5)

# Slow Move

slow_move_folder = local_decrypted_folder+'/slow-move'
slow_moves = process_folder(slow_move_folder, process_slow_move_accelerations_file)
slow_moves_df = pd.DataFrame(slow_moves)
slow_moves_df.to_csv('slow_move_accelerations.csv', index=False)

slow_moves_touches = process_folder(slow_move_folder, process_slow_move_touches_file)
slow_moves_touches_df = pd.DataFrame(slow_moves_touches)
slow_moves_touches_df.to_csv('slow_move_touches.csv', index=False)

participants_with_less_sessions_than_expected(slow_moves_df)
