==> 02_calibration.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from scipy.signal import argrelextrema

from src.accelerometer import plot_fourier_transformation, plot_acceleration
from src.pandas_util import drop_outliers_of_column, split_by_column
from src.plotting import box_plot_columns
from src.time_series import median_filter

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
plt.rcParams.update({'font.size': 18})

calibration_df = pd.read_csv('calibrations.csv')

calibration_df = median_filter(calibration_df)

calibration_df.head(5)

calibration_df.describe()

calibration_df = median_filter(calibration_df)

devices_df = split_by_column(calibration_df,'device')

device_1_df = devices_df[0]
device_2_df = devices_df[1]

single_session_device_1_df = device_1_df[device_1_df['uuid'] == device_1_df[device_1_df['subject'] == '30-02']['uuid'].unique()[3]]
single_session_device_2_df = device_2_df[device_2_df['uuid'] == device_2_df[device_2_df['subject'] == '30-02']['uuid'].unique()[1]]

device_1_hand = device_1_df['hand'].head(1).item()
device_2_hand = device_2_df['hand'].head(1).item()

session_df = [single_session_device_1_df, single_session_device_2_df]

plot_acceleration(session_df,title=[f'Calibration data of a single session of {device_1_hand} hand',f'Calibration data of a single session of {device_2_hand} hand'])

device_1_df.describe()

device_2_df.describe()

def analyse_device_df(df):
    device = df['hand'].unique()[0]
    x_mean = df['x'].mean()
    y_mean = df['y'].mean()
    z_mean = df['z'].mean()
    print(f'device {device} corrections')
    print(f'x correction: {round(x_mean,2)}')
    print(f'y correction: {round(y_mean,2)}')
    print(f'z correction: {round(z_mean,2)}')

for device_df in devices_df:
    analyse_device_df(device_df)
