==> 03_stroop_game_analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from src.stroop_analysis import plot_stroop_stacceleration
from src.accelerometer import plot_fourier_transformation, plot_acceleration, plot_feature_columns, accelerometer_feature_engineering, plot_acceleration_all_subjects
from src.plotting import box_plot_columns
from src.kmeans import kmeans
from src.time_series import median_filter, run_time_series_algorithms
from src.ml_util import run_feature_algorithms
from src.velocity_peaks import velocity_peaks
from src.pandas_util import correlation_matrix, get_min_value_across_columns,extract_sample_sessions,extract_subject_dataframes, reduce_to_one_session_for_subject

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
plt.rcParams.update({'font.size': 18})

stroop_df = pd.read_csv('stroop_accelerations.csv')
stroop_df = stroop_df[stroop_df['age_group']!=0]

stroop_df.head(5)

stroop_df.describe()

#plot_acceleration_all_subjects(stroop_df,"Stroop")

single_session_30_df, single_session_50_df = extract_sample_sessions(stroop_df)
plot_stroop_stacceleration([single_session_30_df, single_session_50_df], ['single session of age-group \'<30\'','single session of age-group \'>50\''])

subject_30_df, subject_50_df = extract_subject_dataframes(stroop_df)
plot_acceleration([subject_30_df,subject_50_df],title=['all sessions of a single participant of age-group \'<30\'','all sessions of a single participant of age-group \'>50\''])

sessions_30_df = reduce_to_one_session_for_subject(stroop_df[stroop_df['age_group']==30])
sessions_50_df = reduce_to_one_session_for_subject(stroop_df[stroop_df['age_group']==50])
plot_acceleration([sessions_30_df,sessions_50_df],title=['one sessions of each participant of age-group \'<30\'','one sessions of each participant of age-group \'>50\''])

plot_fourier_transformation(single_session_30_df, 'test session')

box_plot_columns(single_session_30_df)

stroop_df = median_filter(stroop_df)

stroop_processed_df = accelerometer_feature_engineering(stroop_df)
stroop_processed_df = stroop_processed_df.reset_index(drop=False)

group_by_keys = ['age_group','subject','device', 'hand','uuid']
stroop_data_df = stroop_df.groupby(group_by_keys)[['click_distance_mean','click_distance_std','click_success_rate']].agg('first')
stroop_processed_df = stroop_processed_df.merge(stroop_data_df, on=group_by_keys)

stroop_processed_df.head()

correlation_matrix(stroop_processed_df)

stroop_processed_df = stroop_processed_df.drop(columns=['x_sem', 'y_sem', 'z_sem', 'mag_sem','x_snr','z_snr','y_snr','mag_snr','x_peaks','y_peaks','z_peaks'])

def extract_peaks_for_clicks(base_df, processed_df):
    processed_df_copy = processed_df.copy()
    click_peaks_df = base_df.groupby(['age_group','subject','device', 'hand','uuid','click_number'])[['x', 'y', 'z', 'mag']].agg(velocity_peaks)
    click_peaks_df = click_peaks_df.reset_index(drop=False)
    for uuid in click_peaks_df['uuid'].unique():
        for click_number in click_peaks_df[click_peaks_df['uuid']==uuid]['click_number'].unique():
            session_df = click_peaks_df[(click_peaks_df['uuid'] == uuid) & (click_peaks_df['click_number'] == click_number)]
            for axis in ['x', 'y', 'z', 'mag']:
                column_name = f'click_{click_number}_{axis}_peaks'
                if column_name not in processed_df_copy.columns:
                    processed_df_copy[column_name] =  np.nan
                processed_df_copy.loc[processed_df_copy['uuid'] == uuid, column_name] = session_df[axis].max()
    return processed_df_copy

stroop_processed_df = extract_peaks_for_clicks(stroop_df, stroop_processed_df)

class_key='age_group'
plot_feature_columns(stroop_processed_df,'std')
plot_feature_columns(stroop_processed_df,'mean')
plot_feature_columns(stroop_processed_df,'sal')
box_plot_columns(stroop_processed_df, class_key,['mag_peaks'])
box_plot_columns(stroop_processed_df, class_key,['click_distance_mean','click_distance_std'])
box_plot_columns(stroop_processed_df, class_key,['click_success_rate'])
box_plot_columns(stroop_processed_df, class_key,['duration'])
box_plot_columns(stroop_processed_df, class_key,['click_0_x_peaks', 'click_1_x_peaks', 'click_2_x_peaks', 'click_3_x_peaks', 'click_4_x_peaks'])
box_plot_columns(stroop_processed_df, class_key,['click_0_z_peaks', 'click_1_z_peaks', 'click_2_z_peaks', 'click_3_z_peaks', 'click_4_z_peaks'])
box_plot_columns(stroop_processed_df, class_key,['click_0_y_peaks', 'click_1_y_peaks', 'click_2_y_peaks', 'click_3_y_peaks', 'click_4_y_peaks'])
box_plot_columns(stroop_processed_df, class_key,['click_0_mag_peaks', 'click_1_mag_peaks', 'click_2_mag_peaks', 'click_3_mag_peaks', 'click_4_mag_peaks'])

We can see that there is significant deviation between the two age groups in the features selected below.

feature_keys = ['y_std','y_mean','mag_peaks','duration','hand']

stroop_processed_df = stroop_processed_df.sort_values(by=['age_group', 'subject'])
for feature in feature_keys:
    box_plot_columns(stroop_processed_df, 'subject',[feature],show_column_in_label=False, drop_outlier=False)

# ML Models

results = {}

stroop_processed_df['age_group'].replace(to_replace=30,value=1,inplace=True)
stroop_processed_df['age_group'].replace(to_replace=50,value=0,inplace=True)
stroop_processed_df['hand'].replace(to_replace='dominant',value=1,inplace=True)
stroop_processed_df['hand'].replace(to_replace='non_dominant',value=0,inplace=True)

stroop_df['age_group'].replace(to_replace=30,value=1,inplace=True)
stroop_df['age_group'].replace(to_replace=50,value=0,inplace=True)
stroop_df['hand'].replace(to_replace='dominant',value=1,inplace=True)
stroop_df['hand'].replace(to_replace='non_dominant',value=0,inplace=True)

stroop_processed_df[feature_keys].head()

## Run Algorithms on extracted Features

feature_results = run_feature_algorithms(stroop_processed_df, feature_keys)
results.update(feature_results)

## Run Time Series Algorithm

time_series_results = run_time_series_algorithms(stroop_df)
results.update(time_series_results)

## ML Results

results
