==> 04_spiral_analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from src.accelerometer import plot_fourier_transformation, plot_acceleration, plot_feature_columns, accelerometer_feature_engineering, plot_acceleration_all_subjects
from src.plotting import box_plot_columns
from src.kmeans import kmeans
from src.decision_tree import decision_tree
from src.time_series import median_filter, run_time_series_algorithms
from src.ml_util import run_feature_algorithms
from src.pandas_util import correlation_matrix,extract_sample_sessions,extract_subject_dataframes, reduce_to_one_session_for_subject

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
plt.rcParams.update({'font.size': 18})

spiral_accelerations_df = pd.read_csv('spiral_accelerations.csv')
spiral_accelerations_df = spiral_accelerations_df[spiral_accelerations_df['age_group']!=0]

spiral_drawings_df = pd.read_csv('spiral_drawings.csv')
spiral_drawings_df = spiral_drawings_df[spiral_drawings_df['age_group']!=0]

spiral_accelerations_df.head(3)

spiral_accelerations_df.describe()

spiral_drawings_df.head(3)

spiral_drawings_df.describe()

def clean_accelerations(accelerations_df, drawings_df):
    accelerations_copy_df = accelerations_df.copy()
    for uuid in accelerations_df['uuid'].unique():
        max_duration = drawings_df[drawings_df['uuid'] == uuid]['duration'].max()
        drop_df = accelerations_df[(accelerations_df['uuid'] == uuid) & (accelerations_df['duration'] > max_duration)]
        accelerations_copy_df = accelerations_df.drop(drop_df.index)
    accelerations_copy_df = accelerations_copy_df[accelerations_copy_df['duration']>0]
    return accelerations_copy_df

spiral_accelerations_df = clean_accelerations(spiral_accelerations_df, spiral_drawings_df)

spiral_accelerations_df = median_filter(spiral_accelerations_df)

#plot_acceleration_all_subjects(spiral_accelerations_df,"Spiral")

single_session_30_df, single_session_50_df = extract_sample_sessions(spiral_accelerations_df)
plot_acceleration([single_session_30_df, single_session_50_df], title=['single session of age-group \'<30\'','single session of age-group \'>50\''])

subject_30_df, subject_50_df = extract_subject_dataframes(spiral_accelerations_df)
plot_acceleration([subject_30_df,subject_50_df],title=['all sessions of a single participant of age-group \'<30\'','all sessions of a single participant of age-group \'>50\''])

sessions_30_df = reduce_to_one_session_for_subject(spiral_accelerations_df[spiral_accelerations_df['age_group']==30])
sessions_50_df = reduce_to_one_session_for_subject(spiral_accelerations_df[spiral_accelerations_df['age_group']==50])
plot_acceleration([sessions_30_df,sessions_50_df],title=['one sessions of each participant of age-group \'<30\'','one sessions of each participant of age-group \'>50\''])

plot_fourier_transformation(single_session_30_df, 'test session')

box_plot_columns(single_session_30_df)

plot_acceleration(single_session_30_df)

spiral_accelerations_df = median_filter(spiral_accelerations_df)

spiral_processed_df = accelerometer_feature_engineering(spiral_accelerations_df)

group_by_keys = ['age_group','subject','hand','uuid']
spiral_data_df = spiral_drawings_df.groupby(group_by_keys)[['first_order_smoothness','second_order_smoothness','thightness','zero_crossing_rate']].agg('first')
spiral_processed_df = spiral_processed_df.merge(spiral_data_df, on=group_by_keys)

spiral_processed_df = spiral_processed_df.reset_index(drop=False)

spiral_processed_df.head()

correlation_matrix(spiral_processed_df)

spiral_processed_df = spiral_processed_df.drop(columns=['x_sem', 'y_sem', 'z_sem', 'mag_sem','x_snr', 'y_snr','mag_snr', 'x_peaks','y_peaks','z_peaks'])

class_key = 'age_group'
plot_feature_columns(spiral_processed_df,'std')
plot_feature_columns(spiral_processed_df,'mean')
plot_feature_columns(spiral_processed_df,'sal')
box_plot_columns(spiral_processed_df,class_key, ['z_snr'])
box_plot_columns(spiral_processed_df,class_key, ['mag_peaks'])
box_plot_columns(spiral_processed_df,class_key, ['first_order_smoothness','second_order_smoothness'])
box_plot_columns(spiral_processed_df,class_key, ['thightness'])
box_plot_columns(spiral_processed_df,class_key, ['zero_crossing_rate'])
box_plot_columns(spiral_processed_df,class_key, ['duration'])

# ML Models

results = {}

feature_keys = ['x_std','x_mean','y_mean','z_snr','second_order_smoothness','duration','hand']

spiral_processed_df = spiral_processed_df.sort_values(by=['age_group','subject'])
for feature in feature_keys:
    box_plot_columns(spiral_processed_df, 'subject',[feature],show_column_in_label=False, drop_outlier=False)

spiral_processed_df['age_group'].replace(to_replace=30,value=1,inplace=True)
spiral_processed_df['age_group'].replace(to_replace=50,value=0,inplace=True)
spiral_processed_df['hand'].replace(to_replace='dominant',value=1,inplace=True)
spiral_processed_df['hand'].replace(to_replace='non_dominant',value=0,inplace=True)

spiral_accelerations_df['age_group'].replace(to_replace=30,value=1,inplace=True)
spiral_accelerations_df['age_group'].replace(to_replace=50,value=0,inplace=True)
spiral_accelerations_df['hand'].replace(to_replace='dominant',value=1,inplace=True)
spiral_accelerations_df['hand'].replace(to_replace='non_dominant',value=0,inplace=True)

## Run Algorithms on extracted Features

feature_results = run_feature_algorithms(spiral_processed_df, feature_keys)
results.update(feature_results)

## Run Time Series Algorithm

time_series_results = run_time_series_algorithms(spiral_accelerations_df)
results.update(time_series_results)

## ML Results

results
