==> 05_slow_move_analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from src.accelerometer import plot_fourier_transformation, plot_acceleration, plot_feature_columns, accelerometer_feature_engineering, plot_acceleration_all_subjects
from src.kmeans import kmeans
from src.decision_tree import decision_tree
from src.plotting import box_plot_columns
from src.time_series import median_filter, run_time_series_algorithms
from src.ml_util import run_feature_algorithms
from src.pandas_util import correlation_matrix,extract_sample_sessions,extract_subject_dataframes,reduce_to_one_session_for_subject

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
plt.rcParams.update({'font.size': 18})

slow_move_df = pd.read_csv('slow_move_accelerations.csv')
slow_move_df = slow_move_df[slow_move_df['age_group']!=0]

slow_move_touches_df = pd.read_csv('slow_move_touches.csv')
slow_move_touches_df = slow_move_touches_df[slow_move_touches_df['age_group']!=0]

slow_move_df.head(5)

slow_move_touches_df.head()

slow_move_df.describe()

slow_move_df = median_filter(slow_move_df)

#plot_acceleration_all_subjects(slow_move_df,"Slow move")

single_session_30_df, single_session_50_df = extract_sample_sessions(slow_move_df)
plot_acceleration([single_session_30_df, single_session_50_df], title=['single session of age-group \'<30\'','single session of age-group \'>50\''])

subject_30_df, subject_50_df = extract_subject_dataframes(slow_move_df)
plot_acceleration([subject_30_df,subject_50_df],title=['all sessions of a single participant of age-group \'<30\'','all sessions of a single participant of age-group \'>50\''])

sessions_30_df = reduce_to_one_session_for_subject(slow_move_df[slow_move_df['age_group']==30])
sessions_50_df = reduce_to_one_session_for_subject(slow_move_df[slow_move_df['age_group']==50])
plot_acceleration([sessions_30_df,sessions_50_df],title=['one sessions of each participant of age-group \'<30\'','one sessions of each participant of age-group \'>50\''])

plot_fourier_transformation(single_session_30_df, 'test session')

box_plot_columns(single_session_30_df)

slow_move_processed_df = accelerometer_feature_engineering(slow_move_df)
slow_move_processed_df = slow_move_processed_df.reset_index(drop=False)

slow_move_touches_df['center_line_deviation'] = np.abs(slow_move_touches_df['y'] - slow_move_touches_df['start_y'])

slow_move_processed_df['y_touch_std'] =  np.nan
for uuid in slow_move_touches_df['uuid'].unique():
    current_session = slow_move_touches_df[slow_move_touches_df['uuid']==uuid]
    y_touch_std = current_session['x'].std()
    slow_move_processed_df.loc[slow_move_processed_df['uuid'] == uuid, 'y_touch_std'] = y_touch_std
    slow_move_processed_df.loc[slow_move_processed_df['uuid'] == uuid, 'center_line_deviation'] = current_session['center_line_deviation'].mean()

slow_move_processed_df.head()

correlation_matrix(slow_move_processed_df)

slow_move_processed_df = slow_move_processed_df.drop(columns=['mag_sem','x_snr','y_snr','z_snr','mag_snr','x_peaks','y_peaks','z_peaks'])

plot_feature_columns(slow_move_processed_df,'std')
plot_feature_columns(slow_move_processed_df,'mean')
plot_feature_columns(slow_move_processed_df,'sal')
box_plot_columns(slow_move_processed_df, 'age_group',['x_sem','y_sem','z_sem'])
box_plot_columns(slow_move_processed_df, 'age_group',['mag_peaks'])
box_plot_columns(slow_move_processed_df, 'age_group',['duration'])
box_plot_columns(slow_move_processed_df, 'age_group',['y_touch_std'])
box_plot_columns(slow_move_processed_df, 'age_group',['center_line_deviation'])

# ML Models

results = {} 

feature_keys = ['y_std','x_mean','y_mean','mag_peaks','duration','center_line_deviation','hand']
class_key = ['age_group']

slow_move_processed_df = slow_move_processed_df.sort_values(by=['age_group','subject'])
for feature in feature_keys:
    box_plot_columns(slow_move_processed_df, 'subject',[feature],show_column_in_label=False)

slow_move_processed_df['age_group'].replace(to_replace=30,value=1,inplace=True)
slow_move_processed_df['age_group'].replace(to_replace=50,value=0,inplace=True)
slow_move_processed_df['hand'].replace(to_replace='dominant',value=1,inplace=True)
slow_move_processed_df['hand'].replace(to_replace='non_dominant',value=0,inplace=True)

slow_move_df['age_group'].replace(to_replace=30,value=1,inplace=True)
slow_move_df['age_group'].replace(to_replace=50,value=0,inplace=True)
slow_move_df['hand'].replace(to_replace='dominant',value=1,inplace=True)
slow_move_df['hand'].replace(to_replace='non_dominant',value=0,inplace=True)

## Run Algorithms on extracted Features

feature_results = run_feature_algorithms(slow_move_processed_df, feature_keys)
results.update(feature_results)

## Run Time Series Algorithm

time_series_results = run_time_series_algorithms(slow_move_df)
results.update(time_series_results)

## ML Results

results
